Make the plugin LSP 1.1 compatible
