+++ LSP-metals.sublime-settings
@@ -2,9 +2,16 @@
   // The version of the Metals server artifact. Requires restarting sublime.
   "server_version": "0.9.2",
 
+  // Optional path to the Java home directory. Requires reloading sublime.
+  "java_home": "",
+
+  // Attach to views with base scope source.scala (ST4 only)
+  "selector": "source.scala",
+
   // Optional array of properties to pass along to the Metals server.
   // Example: `-Dhttps.proxyHost=… -Dhttps.proxyPort=…` or `-Dmetals.statistics=all`
   // check https://scalameta.org/metals/docs/editors/new-editor.html#metals-server-properties for all properties
+  // (ST3 only)
   "server_properties": [
     "-Dmetals.client=sublime",
     "-Dmetals.status-bar=on",
@@ -13,6 +20,30 @@
     "-Xms100m"
   ],
 
-  // Optional path to the Java home directory. Requires reloading sublime.
-  "java_home": ""
+  // The startup command (ST4 only)
+  "command": [
+    "${java_path}",
+    "-Dmetals.client=sublime",
+    "-Dmetals.status-bar=on",
+    "-Dmetals.execute-client-command=on",
+    "-Xss4m",
+    "-Xms100m",
+    "-jar",
+    "${packages}/LSP-metals/coursier",
+    "launch",
+    "--ttl",
+    "Inf",
+    "--repository",
+    "bintray:scalacenter/releases",
+    "--repository",
+    "sonatype:snapshots",
+    "--main-class",
+    "scala.meta.metals.Main",
+    "org.scalameta:metals_2.12:${server_version}"
+  ],
+
+  // initialization options (ST4 only)
+  "initializationOptions": {
+    "inputBoxProvider": true
+  }
 }
