+++ common/__init__.py
@@ -1,0 +1,26 @@
+from LSP.plugin.core.types import Optional
+from LSP.plugin.core.url import filename_to_uri
+import os
+import sublime
+import sublime_plugin
+
+
+def get_java_path(settings: sublime.Settings) -> str:
+    java_home = settings.get("java_home")
+    if isinstance(java_home, str) and java_home:
+        return os.path.join(java_home, "bin", "java")
+    java_home = os.environ.get('JAVA_HOME')
+    if java_home:
+        return os.path.join(java_home, "bin", "java")
+    return "java"
+
+
+class NewScalaFileCommand(sublime_plugin.TextCommand):
+
+    def run(self, edit: sublime.Edit, paths) -> None:
+        path = paths[0] if paths else self.view.file_name()
+        args = {
+            "command_name": "new-scala-file",
+            "command_args": [filename_to_uri(path)]
+        }
+        self.view.run_command("lsp_execute", args)
