--- constants.py
@@ -1,3 +1,0 @@
-init_options = {
-    'inputBoxProvider': True
-}
